Fix sync for ProtonVPN OpenVPN configs
ProtonVPN now requires the user to be authenticated to access the
configuration files.

We now request that the user log in via their browser and then copy the
value of the `AUTH-` cookie.

Note the `Feature` specification for filtering for Tor / Torrent
optimised servers has been removed from ProtonVPN's configuration page,
and so is removed here too.

Also note that for SecureCore configs, the first name in the filename is the
output node country, as this is probably the most useful for filtering.

Fixes issue #201