Add port forwarding for ProtonVPN connections
Works with OpenVPN and Wireguard (downloaded ProtonVPN custom config)

Note for OpenVPN port forwarding you must generate the OpenVPN config
files appending "+pmp" to the OpenVPN username.

Note we do not currently handle if the port changes when renewed (i.e.
we do not rest firewall rules in this case).