Fix config parsing issues (#341)
* fix: Protocol in config ignored

Even when Protocol no specified in command line.

Steps to reproduce:

1. echo 'protocol = "Wireguard"' > ~/.config/vopono/config.toml
2. vopono -v exec --provider None bash

Result

 Error: None protocol must use None provider - will run not run any VPN service inside netns

* fix: Don't recreate config file

No need to recreate file, which should be created manually.

Calling with the --vopono-config option also caused the creation of a
new config, but the owner and group of the file were root.

In the end, this also this breaks this workflow:

1. Rename config.toml -> config.toml.bak
2. Rerun vopono
3. Revert renaming

New behavior - empty config builder used if config file doesn't exists,
without empty file creation.

* chore: Generalise config error message

* doc: Update custom option

* fix: Don't error if default config doesn't exists